requirements.txt
flask>=2.0.1
torch>=2.2.0
torchvision>=0.17.0
Pillow>=9.0.0
numpy>=1.21.0